feat: VLAN sub-interface module + PPPoE integration
Major Features:
- Dynamic interface naming based on speed (eth/gi/10g/25g/40g/100g)
- VLAN sub-interface creation from startup.json (eth1.100)
- CLI commands: show vlan, interface eth0.100, no interface eth0.100
- write memory saves VLAN config with type and vlan_id

PPPoE Integration:
- Sessions now use VLAN interface for automatic VLAN tagging
- Removed manual VLAN tagging from ppp_lcp.c, ppp_ipcp.c, ppp_auth.c
- PADS uses interface_send() through VLAN interface

Rate Limiting Fixes (accel-ppp style):
- Increased PADI rate limit from 5/sec to 100/sec per MAC
- Global limit set to 0 (unlimited) like accel-ppp default
- Fixes PPPoE scan disappearing after 2-3 seconds

Files Modified:
- src/interfaces/interface.c: Dynamic naming based on link speed
- src/config/startup_json.c: Brace-depth parsing for interface section
- src/cli/cli_vlan.c: New VLAN CLI module
- src/cli/vty.c, CMakeLists.txt: CLI integration
- src/pppoe/*.c: VLAN interface for sessions, removed manual tagging
- src/pppoe/pppoe_security.c: accel-ppp style rate limiting

+++ src/qos/qos.c
@@ -70,7 +70,8 @@ int qos_tb_conform(struct token_bucket *tb, uint32_t pkt_len)
     tb->last_update = now;
 
     /* Refill tokens: (Bytes/sec * cycles) / cycles/sec */
-    uint64_t new_tokens = (tb->rate * elapsed) / get_hz();
+    /* Use 128-bit math to prevent overflow with high rates (e.g. virtual 100G+) */
+    uint64_t new_tokens = (uint64_t)( ((__uint128_t)tb->rate * elapsed) / get_hz() );
 
     tb->tokens += new_tokens;
     if (tb->tokens > tb->burst) {
@@ -92,7 +93,8 @@ int qos_tb_check(struct token_bucket *tb, uint32_t pkt_len)
     uint64_t elapsed = now - tb->last_update;
     /* Do not update last_update here */
 
-    uint64_t new_tokens = (tb->rate * elapsed) / get_hz();
+    /* Use 128-bit math here too */
+    uint64_t new_tokens = (uint64_t)( ((__uint128_t)tb->rate * elapsed) / get_hz() );
     uint64_t projected_tokens = tb->tokens + new_tokens;
     if (projected_tokens > tb->burst) projected_tokens = tb->burst;
